Snaxi Hotfixes

+++ code/game/turfs/unsimulated/snow.dm
@@ -13,6 +13,8 @@ var/blizzard_ready = 1 //Whether a new blizzard can be started.
 var/list/snowstorm_ambience = list('sound/misc/snowstorm/snowfall_calm.ogg','sound/misc/snowstorm/snowfall_average.ogg','sound/misc/snowstorm/snowfall_hard.ogg','sound/misc/snowstorm/snowfall_blizzard.ogg')
 var/list/snowstorm_ambience_volumes = list(30,40,60,80)
 
+
+
 var/blizzard_cooldown = 3000 //5 minutes minimum
 
 /datum/event/blizzard/start() //Fuck using event code, we'll code all of this here
@@ -21,11 +23,11 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 		command_alert(/datum/command_alert/blizzard_start)
 		sleep(rand(20 SECONDS, 2 MINUTES))
 		greaten_snowfall()
-		sleep(rand(50 SECONDS, 3 MINUTES))
+		sleep(rand(3 MINUTES, 6 MINUTES))
 		greaten_snowfall()
-		sleep(rand(5 MINUTES, 20 MINUTES))
+		sleep(rand(8 MINUTES, 13 MINUTES))
 		lessen_snowfall()
-		sleep(rand(50 SECONDS, 3 MINUTES))
+		sleep(rand(3 MINUTES, 6 MINUTES))
 		lessen_snowfall()
 		sleep(rand(20 SECONDS, 40 SECONDS))
 		command_alert(/datum/command_alert/blizzard_end)
@@ -46,13 +48,13 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 		sleep(rand(3 MINUTES, 5 MINUTES))
 		greaten_snowfall() //Never-ending MISERY
 
-/proc/snowfall_force_update_player() //Since the vision blocking UI only updates on Entered, let's call it.
+/proc/force_update_snowfall_sfx() //Since the vision blocking UI only updates on Entered, let's call it.
 	for(var/mob/M in player_list)
 		if(M && M.client)
 			var/turf/unsimulated/floor/snow/snow = get_turf(M)
 			if(snow && istype(snow))
 				snow.Entered(M)
-				snow.force_update_music(M)
+				M << sound(snowstorm_ambience[snow_intensity+1], repeat = 1, wait = 0, channel = CHANNEL_WEATHER, volume = snowstorm_ambience_volumes[snow_intensity+1])
 
 
 
@@ -65,7 +67,7 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	for(var/turf/unsimulated/floor/snow/tile in global_snowtiles)
 		tile.snow_state = snow_intensity
 		tile.update_environment()
-	snowfall_force_update_player()
+	force_update_snowfall_sfx()
 
 /proc/lessen_snowfall()
 	if(snow_intensity == SNOW_CALM)
@@ -74,7 +76,7 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	for(var/turf/unsimulated/floor/snow/tile in global_snowtiles)
 		tile.snow_state = snow_intensity
 		tile.update_environment()
-	snowfall_force_update_player()
+	force_update_snowfall_sfx()
 
 /proc/snowfall_tick()
 	switch(snow_intensity)
@@ -90,7 +92,7 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 /proc/snowfall_calm_tick()
 	var/tile_interval = 5
 	if(prob(3))
-		var/i = rand(1,tile_interval) //Efficiently selects a set of random tiles to melt snow on.
+		var/i = rand(1,tile_interval) //Efficiently selects a set of random tiles to melt snow on. 
 		for(var/turf/unsimulated/floor/snow/tile in global_snowtiles)
 			if(i == tile_interval)
 				tile.change_snowballs(-1,0)
@@ -154,21 +156,29 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	nitrogen = MOLES_N2STANDARD_ARCTIC
 	can_border_transition = 1
 	plane = PLATING_PLANE
-	var/real_snow_tile = TRUE //Set this to false if you want snowfall/blizzard overlay but no texture updating and ability to pick up snowballs.
+	var/real_snow_tile = TRUE //Set this to false if you want snowfall/blizzard overlay but no texture updating nor ability to pick up snowballs.
 	var/initial_snowballs = -1 //-1 means random.
 	var/snowballs = 0
 	var/snow_state = SNOW_CALM
 	var/obj/effect/snowprint_holder/snowprint_parent
 	var/obj/effect/blizzard_holder/blizzard_parent
 	turf_speed_multiplier = 1
+	gender = PLURAL
+	
+/turf/unsimulated/floor/snow/ChangeTurf(var/turf/N, var/tell_universe=1, var/force_lighting_update = 0, var/allow = 1)
+	if(snowprint_parent)
+		qdel(snowprint_parent)
+	if(blizzard_parent)
+		qdel(blizzard_parent)	
+	..()
 
 /turf/unsimulated/floor/snow/New()
 	..()
 	blizzard_parent = new /obj/effect/blizzard_holder(src)
 	blizzard_parent.parent = src
 	if(!snowtiles_setup)
-		for(var/i = 1 to 4)
-			snow_state = i
+		for(var/i = 0 to 3)
+			snow_state = i 
 			blizzard_parent.UpdateSnowfall()
 		snowtiles_setup = 1
 	snow_state = snow_intensity
@@ -184,8 +194,10 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 
 /turf/unsimulated/floor/snow/Destroy()
 	global_snowtiles -= src
-	qdel(snowprint_parent)
-	qdel(blizzard_parent)
+	if(snowprint_parent)
+		qdel(snowprint_parent)
+	if(blizzard_parent)
+		qdel(blizzard_parent)
 
 /turf/unsimulated/floor/snow/proc/update_environment()
 	if(real_snow_tile)
@@ -201,13 +213,13 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 			temperature = T_ARCTIC
 			turf_speed_multiplier = 1
 		if(SNOW_AVERAGE)
-			temperature = T_ARCTIC-15
+			temperature = T_ARCTIC-5
 			turf_speed_multiplier = 1.15 //For some reason, higher numbers mean slower.
 		if(SNOW_HARD)
-			temperature = T_ARCTIC-30
+			temperature = T_ARCTIC-10
 			turf_speed_multiplier = 1.6
 		if(SNOW_BLIZZARD)
-			temperature = T_ARCTIC-50
+			temperature = T_ARCTIC-20
 			turf_speed_multiplier = 2.9
 	turf_speed_multiplier *= 1+(snowballs/10)
 
@@ -215,20 +227,20 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	..()
 	if(istype(A,/mob/living/carbon/human))
 		var/mob/living/carbon/human/H = A
-		if(snowprint_parent)
+		if(snowprint_parent && snowballs)
 			snowprint_parent.AddSnowprintGoing(H.get_footprint_type(),H.dir)
 		if(!istype(newloc,/turf/unsimulated/floor/snow))
 			H.clear_fullscreen("snowfall_average",0)
 			H.clear_fullscreen("snowfall_hard",0)
 			H.clear_fullscreen("snowfall_blizzard",0)
 			H << sound(null, 0, 0, channel = CHANNEL_WEATHER)
+	
 
-
-/turf/unsimulated/floor/snow/Entered(atom/A, atom/OL)
+/turf/unsimulated/floor/snow/Entered(atom/A, atom/OL)	
 	..()
 	if(istype(A,/mob/living/carbon/human))
 		var/mob/living/carbon/human/H = A
-		if(snowprint_parent)
+		if(snowprint_parent && snowballs)
 			snowprint_parent.AddSnowprintComing(H.get_footprint_type(),H.dir)
 		switch(snow_state)
 			if(SNOW_CALM)
@@ -252,21 +264,18 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 			if(isliving(H) && !H.locked_to && !H.lying && !H.flying)
 				playsound(src, pick(snowsound), 10, 1, -1, channel = 123)
 
+/turf/unsimulated/floor/snow/cultify()
+	return //It's already pretty red out in nar-sie universe.
 
-/turf/unsimulated/floor/snow/proc/force_update_music(var/mob/M)
-	if(M.client)
-		M << sound(snowstorm_ambience[snow_state+1], repeat = 1, wait = 0, channel = CHANNEL_WEATHER, volume = snowstorm_ambience_volumes[snow_state+1])
-
-
-/obj/effect/blizzard_holder //Exists to make it unclickable
+/obj/effect/blizzard_holder //Exists to make it unclickable 
 	name = "blizzard"
 	desc = "Brrr."
 	density = 0
 	anchored = 1
-	plane = ABOVE_TURF_PLANE
-	mouse_opacity = 0
+	plane = ABOVE_TURF_PLANE	
+	mouse_opacity = 0 
 	var/turf/unsimulated/floor/snow/parent
-
+	
 /obj/effect/blizzard_holder/proc/UpdateSnowfall()
 	if(!snow_state_to_texture["[parent.snow_state]"])
 		cache_snowtile()
@@ -279,9 +288,8 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	for(var/i = 1 to overlay_counts[parent.snow_state+1])
 		var/image/snowfx = image('icons/turf/snowfx.dmi', "[snowfall_overlays[parent.snow_state+1]][i]",SNOW_OVERLAY_LAYER)
 		snowfx.plane = EFFECTS_PLANE
-		overlays += snowfx
+		overlays += snowfx			
 	snow_state_to_texture["[parent.snow_state]"] = appearance
-
 
 
 
@@ -294,7 +302,7 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 	mouse_opacity = 0 //Unclickable
 	var/snowprint_color = "#BEBEBE"
 	var/list/existing_prints = list()
-
+	
 /obj/effect/snowprint_holder/proc/AddSnowprintComing(var/obj/effect/decal/cleanable/blood/tracks/footprints/footprint_type, var/dir)
 	if(existing_prints["[initial(footprint_type.coming_state)]-[dir]"])
 		return
@@ -411,12 +419,43 @@ var/blizzard_cooldown = 3000 //5 minutes minimum
 		return BUILD_SUCCESS
 	return BUILD_FAILURE
 
+	
+	
+	
+/turf/unsimulated/floor/snow/asphalt
+	icon = 'icons/turf/floors.dmi'
+	icon_state = "concrete"
+	real_snow_tile = FALSE
+	name = "asphalt"
+	desc = "Specially treated Centcomm asphalt, designed to disintegrate all snow that touches it."	
+	
 /turf/unsimulated/floor/snow/permafrost
 	icon_state = "permafrost_full"
 	real_snow_tile = FALSE
 	name = "permafrost"
 	desc = "Soil that never unfreezes."
-
+	
+/turf/unsimulated/floor/noblizz_permafrost
+	icon = 'icons/turf/new_snow.dmi'
+	icon_state = "permafrost_full"
+	name = "permafrost"
+	desc = "Soil that never unfreezes."
+	gender = PLURAL
+	temperature = T_ARCTIC
+	oxygen = MOLES_O2STANDARD_ARCTIC
+	nitrogen = MOLES_N2STANDARD_ARCTIC
+	can_border_transition = 1
+	plane = PLATING_PLANE
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 /obj/glacier
 	desc = "A frozen lake kept solid by temperatures way below freezing."
 	icon = 'icons/turf/ice.dmi'
